Display category description according to url and rename guitar_list
The guitar_list.ejs file was ranamed to item_list.ejs, as the template
will be used to display the information about the remaining categories.

On the other hand, the name of the category is taken from the url by
setting the :name parameter, which will be used to extract the data
of the respective category by finding one element by its name.